Fix spell check ~false positive by ignoring word (#22)
In the latest release of the codespell tool used for automated spell checking of the files of this project, the word "inflight" was added to the codespell misspelled words dictionary as a misspelling of "in-flight".

The "inflight" spelling occurs in a comment in `src/utility/MQTTClient.h`, which caused the spell check to start failing.

Although that is surely correct for the prose word, "inflight" seems to be somewhat common in the context of MQTT, and also occurs in names in the code, so I'm not sure that "correcting" the current spelling would be the right thing to do. It is also probably questionable whether this file should be edited at all, since it appears it might be an externally maintained resource (meaning that any changes to it in this repo make a hypothetical sync from upstream more difficult).

All in all, it seems best to simply configure codespell to ignore the problematic word.