Correct typos in comments
The codespell spellchecker tool is used to automatically detect commonly misspelled words in the files of this project.

The misspelled words dictionary was expanded in the latest release of codespell, which resulted in the detection of
misspelled words in the project's comments.

The typos are hereby corrected, which will restore the spell check to a passing state.