Use standardized `maintainer` metadata property value
The `maintainer` property of the library.properties metadata file defines the entity responsible for maintenance of the
library.

In the libraries that are under Arduino's maintainership, it is standard practice to set the `maintainer` property to
the general purpose value "Arduino <info@arduino.cc>".

Previously, instead of that standard value, this library's `maintainer` property defined a former partner company which
hasn't had any involvement in the library for many years, and is unlikely to do so at any point in the future. So the
metadata is changed to use the standard value.